requirements.txt
torch==2.2.1
torchvision
torchaudio
transformers==4.51.3
Pillow